requirements.txt
psycopg2-binary==2.9.10
ollama==0.3.3
Flask==3.0.3
python-dotenv==1.0.1
